requirement.txt
python-telegram-bot==20.6
python-dotenv
requests
beautifulsoup4
selenium
networkx
matplotlib
pandas
lxml
